hehe
Can scrape every series and every episode. Running the tester application returns a /hh/ directory which contains every series with description and all episode direct links + the thumbnails. Artificial wait in between the scrapes to not get banned from requesting. (to minimize requests, do "GetSeriesList()" once and save it to "series.txt" so you dont have to request the list all the time.